requirements.txt
# Core ML & Data Processing
torch==2.1.0
torchvision==0.16.0
fastai==2.7.13
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
scikit-learn==1.3.1
opencv-python==4.8.1.78

# MLflow & Experiment Tracking
mlflow==2.8.0
boto3==1.28.57

# API & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
streamlit==1.28.1
gradio==4.4.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==1.4.50
alembic==1.12.1

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-amazon==8.9.0

# Object Storage (S3/MinIO)
minio==7.2.0

# Monitoring
prometheus-client==0.18.0
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
loguru==0.7.2

# Feature Store
# feast==0.35.0  # Temporairement désactivé - conflit avec Pydantic 2

# Load Testing
locust==2.17.0
